machinelearning/requirements.txt
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Flask==2.2.3
fonttools==4.39.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
liac-arff==2.5.0
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
threadpoolctl==3.1.0
Werkzeug==2.2.3
zipp==3.15.0
